qrcode expire

+++ src/views/deveops/manager/host.vue
@@ -15,8 +15,7 @@
           inactive-text="详细检索">
         </el-switch>
         <el-button class="filter-item" @click="resetSearch()" style="margin-left: 10px;" type="primary" icon="el-icon-refresh" :disabled="btnStatus">清除</el-button>
-        <el-button class="filter-item" @click="handleMultipleGroup()" style="margin-left: 10px;" type="primary" icon="el-icon-goods" :disabled="btnStatus">归类</el-button>
-        <el-button class="filter-item" @click="handleExpired()" style="margin-left: 10px;" type="primary" icon="el-icon-time" :disabled="btnStatus">过期资源</el-button>
+        <el-button class="filter-item" @click="handleSelectHost()" style="margin-left: 10px;" type="primary" icon="el-icon-goods" :disabled="btnStatus">归类</el-button>
         <el-button class="filter-item" @click="handleCreate()" style="float:right;" type="primary" icon="el-icon-edit" :disabled="btnStatus">新增</el-button>
       </el-row>
       <el-row>
@@ -98,7 +97,7 @@
         <template slot-scope="host">
           <!-- <el-button type="primary" size="medium" @click="handleQRCode(host.row)" :disabled="btnStatus">密码</el-button> -->
           <el-button type="primary" size="medium" @click="handleDetail(host.row)" :disabled="btnStatus">详细</el-button>
-          <el-button type="warning" size="medium" @click="handleGroup(host.row)" :disabled="btnStatus">应用组</el-button>
+          <el-button type="warning" size="medium" @click="handleSelectGroup(host.row)" :disabled="btnStatus">应用组</el-button>
           <el-button type="warning" size="medium" @click="handleUpdate(host.row)" :disabled="btnStatus">编辑</el-button>
           <el-button type="danger" size="medium" @click="handleDelete(host.row)" :disabled="btnStatus">删除</el-button>
         </template>
@@ -110,8 +109,7 @@
       @size-change="handleSizeChange"
       @current-change="handleCurrentChange" 
       :total="pagination.count"
-      :page-sizes="[10, 20, 30, 40, 50, 60]"
-      >
+      :page-sizes="[10, 20, 30, 40, 50, 60]">
       </el-pagination>
     </div>
 
@@ -126,19 +124,19 @@
       </el-select>
       <div slot="footer" class="dialog-footer">
         <el-button @click="dialogSelectHostVisible = false" :disabled="btnStatus">取消</el-button>
-        <el-button type="primary" @click="selectGroup" :disabled="btnStatus">提交</el-button>
+        <el-button type="primary" @click="handleQRCode" :disabled="btnStatus">提交</el-button>
       </div>
     </el-dialog>
 
-    <el-dialog :title="textMap[dialogStatus]" :visible.sync="dialogPasswdVisible" width="20%" top="20vh">
+    <el-dialog title="QRCode二次验证" :visible.sync="dialogQRCodeVisible" width="30%" top="20vh">
       <span>请确认您的权限是运维工程师并且已经拥有QR-Code</span>
       <el-input v-model="commit_obj.qrcode" placeholder="请输入您当前账户的QR-Code"></el-input>
-
       <div slot="footer" class="dialog-footer">
-        <el-button type="primary" icon="document" 
-        @click="UpdateQRCode">校验QR-Code
-      </el-button>
-        <el-button v-show="temp_passwd" type="primary" icon="document" v-clipboard:copy='temp_passwd' v-clipboard:success='clipboardSuccess' @click="dialogPasswdVisible = false">粘贴密码</el-button>
+        <el-button v-if="dialogStatus=='create'" type="primary" @click="createData" :disabled="btnStatus">创建</el-button>
+        <el-button v-else-if="dialogStatus=='update'" type="primary" @click="updateData" :disabled="btnStatus">更新</el-button>
+        <el-button v-else-if="dialogStatus=='delete'" type="primary" @click="deleteConfirm" :disabled="btnStatus">删除</el-button>
+        <el-button v-else-if="dialogStatus=='selecthost'" type="primary" @click="selectHost" :disabled="btnStatus">批量归类权限</el-button>
+        <el-button v-else-if="dialogStatus=='selectgroup'" type="primary" @click="selectGroup" :disabled="btnStatus">归类权限</el-button>
       </div>
     </el-dialog>
 
@@ -224,7 +222,7 @@
       </div>
     </el-dialog>
 
-    <el-dialog :title="textMap[dialogStatus]" :visible.sync="dialogGroupVisible" width="60%" top="2vh">
+    <el-dialog :title="textMap[dialogStatus]" :visible.sync="dialogSelectGroupVisible" width="60%" top="2vh">
       <el-form ref="groupForm" :model="commit_obj" label-position="left" label-width="100px" style='width: 700px; margin-left:40px;'>
         <el-form-item label="所属权限组" prop="pmn_groups">
           <el-transfer v-model="commit_obj.groups" :data="groups" placeholder="请选择所属应用组" filterable>
@@ -232,8 +230,8 @@
         </el-form-item>
       </el-form>
       <div slot="footer" class="dialog-footer">
-        <el-button @click="dialogGroupVisible = false" :disabled="btnStatus">取消</el-button>
-        <el-button type="primary" @click="updateGroups" :disabled="btnStatus">提交</el-button>
+        <el-button @click="dialogSelectGroupVisible = false" :disabled="btnStatus">取消</el-button>
+        <el-button type="primary" @click="handleQRCode" :disabled="btnStatus">提交</el-button>
       </div>
     </el-dialog>
 
@@ -306,8 +304,7 @@
       </el-form>
       <div slot="footer" class="dialog-footer">
         <el-button @click="dialogFormVisible = false" :disabled="btnStatus">取消</el-button>
-        <el-button v-if="dialogStatus=='create'" type="primary" @click="createData" :disabled="btnStatus">提交</el-button>
-        <el-button v-else type="primary" @click="updateData" :disabled="btnStatus">提交</el-button>
+        <el-button type="primary" @click="handleQRCode" :disabled="btnStatus">提交</el-button>
       </div>
 
     </el-dialog>
@@ -316,11 +313,12 @@
 </template>
 
 <script>
-import IEcharts from "vue-echarts-v3/src/lite.js";
-import "echarts/lib/chart/bar";
-import "echarts/lib/chart/line";
-import "echarts/lib/component/title";
-import echarts from "echarts";
+import IEcharts from "vue-echarts-v3/src/lite.js"
+import "echarts/lib/chart/bar"
+import "echarts/lib/chart/line"
+import "echarts/lib/component/title"
+import echarts from "echarts"
+import { is_expire_User } from '@/api/auth'
 import {
   fetch_HostListByPage,
   delete_Host,
@@ -348,14 +346,11 @@ export default {
       btnStatus: false,
       dialogDetailVisible: false,
       dialogFormVisible: false,
-      dialogPasswdVisible: false,
-      dialogGroupVisible: false,
       dialogSelectHostVisible: false,
+      dialogSelectGroupVisible: false,
+      dialogQRCodeVisible: false,
       detailSearch: false,
-      systemtype: [],
-      position: [],
       monitorFlag: null,
-      temp_passwd: "",
       groups: [],
       detail_time: 1,
       multipleSelection: [],
@@ -370,7 +365,8 @@ export default {
         create: "新建主机",
         passwd: "粘贴密码",
         group: "修改应用组",
-        selecthost: "批量归类主机"
+        selecthost: "批量归类主机",
+        selectgroup: "归类主机"
       },
       dialogStatus: "",
       monitor_obj: {},
@@ -501,13 +497,13 @@ export default {
       ],
       rules: {
         info: [
-          { required: true, message: "主机信息是必须的", trigger: "change" }
+          { required: true, message: "主机信息是必须的", trigger: "blur" }
         ],
         connect_ip: [
           {
             required: true,
             message: "连接IP是您管理主机的重要信息",
-            trigger: "change"
+            trigger: "blur"
           },
           {
             pattern: /^(\d{1,2}|1\d\d|2[0-4]\d|25[0-5])(\.(\d{1,2}|1\d\d|2[0-4]\d|25[0-5])){3}$/,
@@ -519,7 +515,7 @@ export default {
           {
             required: true,
             message: "连接端口是您管理主机的重要信息",
-            trigger: "change"
+            trigger: "blur"
           }
         ],
         _status: [
@@ -670,12 +666,12 @@ export default {
       });
     },
     handleSizeChange(val) {
-      this.pagination.page_size = val;
-      this.init_hosts();
+      this.pagination.page_size = val
+      this.init_hosts()
     },
     handleCurrentChange(val) {
-      this.pagination.page = val;
-      this.init_hosts();
+      this.pagination.page = val
+      this.init_hosts()
     },
     tableRowClassName({ row, rowIndex }) {
       if (row._status === -3) {
@@ -690,10 +686,17 @@ export default {
       }
     },
     reset_commit() {
-      this.commit_obj = {};
+      this.commit_obj = {}
     },
     reset_search() {
-      this.search_obj = {};
+      this.search_obj = {}
+    },
+    reset_dialog() {
+      this.dialogSelectHostVisible = false
+      this.dialogSelectGroupVisible = false
+      this.dialogFormVisible = false
+      this.dialogDetailVisible = false
+      this.dialogQRCodeVisible = false
     },
     changeGroup() {
       this.pagination = {
@@ -701,23 +704,23 @@ export default {
         count: 0,
         page_size: 10
       };
-      this.init_hosts();
+      this.init_hosts()
     },
     clearGroup() {
       this.pagination = {
         page: 1,
         count: 0,
         page_size: 10
-      };
-      this.init_hosts();
+      }
+      this.init_hosts()
     },
     searchHost() {
       this.pagination = {
         page: 1,
         count: 0,
         page_size: 10
-      };
-      this.init_hosts();
+      }
+      this.init_hosts()
     },
     handleSelectionChange(val) {
       this.multipleSelection = val;
@@ -734,7 +737,8 @@ export default {
             message: "删除成功",
             type: "success"
           });
-          this.init();
+          this.reset_dialog()
+          this.init()
         });
       });
     },
@@ -743,21 +747,21 @@ export default {
       this.dialogStatus = "detail";
       this.monitorFlag = 0;
       this.detail_time = 1;
-      this.init_detail_aliyun();
+      this.init_detail_aliyun()
     },
     handleCreate(row) {
-      this.reset_commit();
-      this.dialogStatus = "create";
-      this.dialogFormVisible = true;
+      this.reset_commit()
+      this.dialogStatus = "create"
+      this.dialogFormVisible = true
       this.$nextTick(() => {
-        this.$refs["dataForm"].clearValidate();
-      });
+        this.$refs["dataForm"].clearValidate()
+      })
     },
     resetSearch() {
-      this.reset_search();
-      this.init_hosts();
+      this.reset_search()
+      this.init_hosts()
     },
-    handleMultipleGroup() {
+    handleSelectHost() {
       if (this.multipleSelection.length == 0) {
         this.$message({
           type: "error",
@@ -765,11 +769,11 @@ export default {
         });
         return;
       }
-      this.reset_commit();
-      this.dialogSelectHostVisible = true;
-      this.dialogStatus = "selecthost";
+      this.reset_commit()
+      this.dialogSelectHostVisible = true
+      this.dialogStatus = "selecthost"
     },
-    selectGroup() {
+    selectHost() {
       let list = [];
       for (const select of this.multipleSelection) {
         list.push(select.id);
@@ -781,8 +785,8 @@ export default {
             message: "归类成功",
             type: "success"
           });
-          this.reset_commit();
-          this.init();
+          this.reset_commit()
+          this.init()
         })
         .catch(error => {
           this.$message({
@@ -791,15 +795,15 @@ export default {
             type: "danger"
           });
         });
-      this.dialogSelectHostVisible = false;
+      this.reset_dialog()
     },
-    handleGroup(row) {
-      this.commit_obj = Object.assign({}, row); // copy obj
-      this.dialogStatus = "group";
-      this.dialogGroupVisible = true;
+    handleSelectGroup(row) {
+      this.commit_obj = Object.assign({}, row) // copy obj
+      this.dialogStatus = "selectgroup"
+      this.dialogSelectGroupVisible = true
       this.$nextTick(() => {
-        this.$refs["groupForm"].clearValidate();
-      });
+        this.$refs["groupForm"].clearValidate()
+      })
     },
     handleUpdate(row) {
       this.commit_obj = Object.assign({}, row); // copy obj
@@ -810,44 +814,45 @@ export default {
         this.$refs["dataForm"].clearValidate();
       });
     },
-    handleQRCode(row) {
-      this.dialogStatus = "passwd";
-      this.dialogPasswdVisible = true;
-      this.commit_obj.uuid = row.uuid;
-    },
-    UpdateQRCode() {
-      fetch_HostPasswd(this.commit_obj).then(response => {
-        this.$message({
-          message: "校验成功",
-          type: "success",
-          duration: 1500
-        });
-        this.temp_passwd = response.data[0].passwd;
-      });
-    },
-    clipboardSuccess() {
-      this.reset_commit();
-      this.temp_passwd = "";
-      this.dialogPasswdVisible = false;
-      this.$message({
-        message: "复制成功",
-        type: "success",
-        duration: 1500
-      });
+    handleQRCode() {
+      is_expire_User()
+        .then((response) => {
+          if (response.data.isexpire) {
+            this.reset_dialog()
+            this.dialogQRCodeVisible = true
+          } else {
+            if (this.dialogStatus === "create") {
+              this.createData()
+            } else if (this.dialogStatus === "update") {
+              this.updateData()
+            } else if (this.dialogStatus === "selecthost"){
+              this.selectHost()
+            } else if (this.dialogStatus === 'selectgroup'){
+              this.selectGroup()
+            } else if (this.dialogStatus === "delete"){
+              this.deleteConfirm()
+            }
+          }
+        }).catch((error) => {
+          console.log(error)
+          this.$message({
+            showClose: true,
+            message: "过期时间确定失败",
+            type: "danger"
+          })
+        })
     },
     handleDelete(row) {
       this.commit_obj = Object.assign({}, row);
-      this.btnStatus = true;
-      this.deleteConfirm();
-      this.btnStatus = false;
+      this.dialogStatus = 'delete'
+      this.handleQRCode()
     },
-    updateGroups(row) {
+    selectGroup(row) {
       this.$refs["groupForm"].validate(valid => {
         if (valid) {
           this.btnStatus = true;
           selectGroup_Host(this.commit_obj)
             .then(() => {
-              this.dialogGroupVisible = false;
               this.init();
               this.$message({
                 showClose: true,
@@ -857,11 +862,11 @@ export default {
               this.btnStatus = false;
             })
             .catch(error => {
-              this.dialogGroupVisible = false;
               this.btnStatus = false;
-            });
+            })
         }
-      });
+        this.reset_dialog()
+      })
     },
     createData() {
       this.$refs["dataForm"].validate(valid => {
@@ -883,6 +888,7 @@ export default {
               this.dialogFormVisible = false;
             });
         }
+        this.reset_dialog()
       });
     },
     updateData() {
@@ -903,34 +909,11 @@ export default {
             .catch(error => {
               this.btnStatus = false;
               this.dialogFormVisible = false;
-            });
+            })
         }
-      });
-    },
-    handleExpired() {
-      this.$router.push({ path: "/manager/expired" });
+        this.reset_dialog()
+      })
     }
-
-    // createSystypeFilter(queryString) {
-    //   return (systemtype) => {
-    //     return (systemtype.name.toLowerCase().indexOf(queryString.toLowerCase()) === 0);
-    //   };
-    // },
-    // createPositionFilter(queryString) {
-    //   return (position) => {
-    //     return (position.name.toLowerCase().indexOf(queryString.toLowerCase()) === 0);
-    //   };
-    // },
-    // queryPositionSearch(queryString,cb) {
-    //   const restaurants = this.position
-    //   const results = queryString ? restaurants.filter(this.createPositionFilter(queryString)) : restaurants
-    //   cb(results)
-    // },
-    // querySystypeSearch(queryString, cb) {
-    //   const restaurants = this.systemtype
-    //   const results = queryString ? restaurants.filter(this.createSystypeFilter(queryString)) : restaurants
-    //   cb(results)
-    // }
   }
 };
 </script>
